Update README.md
I am not sure if you are having 3rd party explanations listed here, but this might be useful as an entry point in your documentation for other reviews and explanations of this plugin. I won't be offended if you don't want to include any of these suggestions, but I personally think its a good idea when plugin docs include 3rd party explanations.